+++ changes.md
@@ -1,7 +1,7 @@
 # Morbid
 
  - Add endpoint to get admin users by accounts
-   - `GET /service/app/{app}/accounts/adms`
+   - `GET /service/app/{app}/accounts/admins`
  - Implementation to call service routes for MorbidClient (client.scala)
 
 ## Release v2.7.0